chore: directory structure
