+++ src/components/ComingSoon.tsx
@@ -1,0 +1,5 @@
+import { FunctionComponent } from 'react';
+
+export function ComingSoon() {
+  return (() => <span>Coming Soon...</span>) as FunctionComponent;
+}
