+++ src/config/bridge/bridge.ts
@@ -1,14 +1,8 @@
 import { isEqual } from 'lodash';
 import { FunctionComponent } from 'react';
 import { ApiKeys, BridgeConfig, ContractConfig, Departure } from '../../model';
-import { darwiniaCrabDVMConfig } from './darwinia-crabDVM';
-import { ethereumCrabDVMConfig } from './ethereum-crabDVM';
-import { ethereumDarwiniaConfig } from './ethereum-darwinia';
-import { pangoroPangolinDVMConfig } from './pangoro-pangolinDVM';
-import { ropstenPangolinConfig } from './ropsten-pangolin';
-import { ropstenPangolinDVMConfig } from './ropsten-pangolinDVM';
 
-type BridgeStatus = 'pending' | 'available';
+export type BridgeStatus = 'pending' | 'available';
 
 /**
  * departure -> arrival: issuing;
@@ -87,57 +81,3 @@ export class Bridge<C extends ContractConfig = ContractConfig, A extends ApiKeys
     return this.record.get(this.redeem);
   }
 }
-
-/**
- * ethereum <-> darwinia
- */
-export const ethereumDarwinia = new Bridge(
-  { network: 'ethereum', mode: 'native' },
-  { network: 'darwinia', mode: 'native' },
-  ethereumDarwiniaConfig
-);
-
-export const ropstenPangolin = new Bridge(
-  { network: 'ropsten', mode: 'native' },
-  { network: 'pangolin', mode: 'native' },
-  ropstenPangolinConfig
-);
-
-/**
- * substrate <-> substrate dvm
- */
-export const darwiniaCrabDVM = new Bridge(
-  { network: 'darwinia', mode: 'native' },
-  { network: 'crab', mode: 'dvm' },
-  darwiniaCrabDVMConfig
-);
-
-export const pangoroPangolinDVM = new Bridge(
-  { network: 'pangoro', mode: 'native' },
-  { network: 'pangolin', mode: 'dvm' },
-  pangoroPangolinDVMConfig
-);
-
-/**
- * ethereum <-> crab dvm
- */
-export const ethereumCrabDVM = new Bridge(
-  { network: 'ethereum', mode: 'native' },
-  { network: 'crab', mode: 'dvm' },
-  ethereumCrabDVMConfig
-);
-
-export const ropstenPangolinDVM = new Bridge(
-  { network: 'ropsten', mode: 'native' },
-  { network: 'pangolin', mode: 'dvm' },
-  ropstenPangolinDVMConfig
-);
-
-export const BRIDGES = [
-  ethereumDarwinia,
-  ropstenPangolin,
-  darwiniaCrabDVM,
-  pangoroPangolinDVM,
-  ethereumCrabDVM,
-  ropstenPangolinDVM,
-];
