+++ src/config/bridge/bridges.ts
@@ -1,0 +1,15 @@
+import { darwiniaCrabDVM } from './darwinia-crabDVM';
+import { ethereumCrabDVM } from './ethereum-crabDVM';
+import { ethereumDarwinia } from './ethereum-darwinia';
+import { pangoroPangolinDVM } from './pangoro-pangolinDVM';
+import { ropstenPangolin } from './ropsten-pangolin';
+import { ropstenPangolinDVM } from './ropsten-pangolinDVM';
+
+export const BRIDGES = [
+  darwiniaCrabDVM,
+  ethereumCrabDVM,
+  ethereumDarwinia,
+  pangoroPangolinDVM,
+  ropstenPangolin,
+  ropstenPangolinDVM,
+];
