+++ src/config/bridge/ethereum-crabDVM.ts
@@ -1,7 +1,8 @@
 import { EthereumCrabDVMConfig } from '../../model';
 import { EVOLUTION_DOMAIN } from '../network';
+import { Bridge } from './bridge';
 
-export const ethereumCrabDVMConfig: EthereumCrabDVMConfig = {
+const ethereumCrabDVMConfig: EthereumCrabDVMConfig = {
   specVersion: 1180,
   api: { dapp: 'https://api.darwinia.network', evolution: EVOLUTION_DOMAIN.product },
   contracts: {
@@ -9,3 +10,9 @@ export const ethereumCrabDVMConfig: EthereumCrabDVMConfig = {
     redeem: '',
   },
 };
+
+export const ethereumCrabDVM = new Bridge(
+  { network: 'ethereum', mode: 'native' },
+  { network: 'crab', mode: 'dvm' },
+  ethereumCrabDVMConfig
+);
