+++ src/config/bridge/index.ts
@@ -1,0 +1,8 @@
+export * from './bridge';
+export * from './bridges';
+export * from './darwinia-crabDVM';
+export * from './ethereum-crabDVM';
+export * from './ethereum-darwinia';
+export * from './pangoro-pangolinDVM';
+export * from './ropsten-pangolin';
+export * from './ropsten-pangolinDVM';
