+++ src/config/bridge/ropsten-pangolinDVM.ts
@@ -1,7 +1,8 @@
 import { EthereumCrabDVMConfig } from '../../model';
 import { EVOLUTION_DOMAIN } from '../network';
+import { Bridge } from './bridge';
 
-export const ropstenPangolinDVMConfig: EthereumCrabDVMConfig = {
+const ropstenPangolinDVMConfig: EthereumCrabDVMConfig = {
   specVersion: 27020,
   api: { dapp: 'https://api.darwinia.network', evolution: EVOLUTION_DOMAIN.product },
   contracts: {
@@ -9,3 +10,9 @@ export const ropstenPangolinDVMConfig: EthereumCrabDVMConfig = {
     redeem: '',
   },
 };
+
+export const ropstenPangolinDVM = new Bridge(
+  { network: 'ropsten', mode: 'native' },
+  { network: 'pangolin', mode: 'dvm' },
+  ropstenPangolinDVMConfig
+);
