+++ src/config/index.ts
@@ -1,5 +1,5 @@
 export * from './abi';
-export * from './birdge';
+export * from './bridge';
 export * from './constant';
 export * from './darwinia';
 export * from './erc20';
