+++ src/utils/index.ts
@@ -1,4 +1,5 @@
 export * from './airdrop';
+export * from './bridge';
 export * from './helper';
 export * from './mmr';
 export * from './network';
